requirements.txt
aiohttp==3.8.4
asyncio==3.4.3
DateTime==5.1
jsonschema==4.17.3
pandas==1.5.3
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
